+++ src/main/java/br/com/zalf/prolog/webservice/gente/prontuarioCondutor/ProntuarioCondutorResource.java
@@ -1,15 +1,16 @@
 package br.com.zalf.prolog.webservice.gente.prontuarioCondutor;
 
 import br.com.zalf.prolog.webservice.commons.network.Response;
-import br.com.zalf.prolog.webservice.commons.util.date.Now;
 import br.com.zalf.prolog.webservice.commons.util.Platform;
 import br.com.zalf.prolog.webservice.commons.util.UsedBy;
+import br.com.zalf.prolog.webservice.commons.util.date.Now;
 import br.com.zalf.prolog.webservice.gente.prontuarioCondutor.model.ProntuarioCondutor;
 import br.com.zalf.prolog.webservice.interceptors.auth.Secured;
 import br.com.zalf.prolog.webservice.permissao.pilares.Pilares;
 import org.apache.commons.io.IOUtils;
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 import org.glassfish.jersey.media.multipart.FormDataParam;
+import org.jetbrains.annotations.NotNull;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -25,20 +26,25 @@ import java.util.List;
 @Path("/prontuarios")
 @Consumes(MediaType.APPLICATION_JSON + ";charset=utf-8")
 @Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
-public class ProntuarioCondutorResource {
+public final class ProntuarioCondutorResource {
 
-    private ProntuarioCondutorService service = new ProntuarioCondutorService();
+    @NotNull
+    private final ProntuarioCondutorService service = new ProntuarioCondutorService();
 
     @GET
     @Path("/{cpf}")
-    @Secured(permissions = {Pilares.Gente.ProntuarioCondutor.VISUALIZAR_PROPRIO, Pilares.Gente.ProntuarioCondutor.VISUALIZAR_TODOS})
+    @Secured(permissions = {
+            Pilares.Gente.ProntuarioCondutor.VISUALIZAR_PROPRIO,
+            Pilares.Gente.ProntuarioCondutor.VISUALIZAR_TODOS})
     public ProntuarioCondutor getProntuario(@PathParam("cpf") Long cpf) {
         return service.getProntuario(cpf);
     }
 
     @GET
     @Path("/{codUnidade}/{codEquipe}")
-    @Secured(permissions = {Pilares.Gente.ProntuarioCondutor.VISUALIZAR_TODOS, Pilares.Gente.ProntuarioCondutor.UPLOAD})
+    @Secured(permissions = {
+            Pilares.Gente.ProntuarioCondutor.VISUALIZAR_TODOS,
+            Pilares.Gente.ProntuarioCondutor.UPLOAD})
     public List<ProntuarioCondutor> getResumoProntuarios(@PathParam("codUnidade") Long codUnidade,
                                                          @PathParam("codEquipe") String codEquipe) {
         return service.getResumoProntuarios(codUnidade, codEquipe);
@@ -46,7 +52,9 @@ public class ProntuarioCondutorResource {
 
     @GET
     @Path("/{cpf}/pontuacao-total")
-    @Secured(permissions = {Pilares.Gente.ProntuarioCondutor.VISUALIZAR_PROPRIO, Pilares.Gente.ProntuarioCondutor.VISUALIZAR_TODOS})
+    @Secured(permissions = {
+            Pilares.Gente.ProntuarioCondutor.VISUALIZAR_PROPRIO,
+            Pilares.Gente.ProntuarioCondutor.VISUALIZAR_TODOS})
     public Double getPontuacaoProntuario(@PathParam("cpf") Long cpf) {
         return service.getPontuacaoProntuario(cpf);
     }
@@ -61,21 +69,22 @@ public class ProntuarioCondutorResource {
             @FormDataParam("file") FormDataContentDisposition fileDetail,
             @PathParam("codUnidade") Long codUnidade) {
         try {
-            String fileName = String.valueOf(Now.utcMillis()) + "_" +
-                    String.valueOf(codUnidade) + "_" + fileDetail.getFileName().replace(" ", "_");
-            File tmpDir = new File(System.getProperty("java.io.tmpdir"), "mapas");
+            final String fileName = Now.utcMillis() + "_" +
+                    codUnidade + "_" + fileDetail.getFileName().replace(" ", "_");
+            final File tmpDir = new File(System.getProperty("java.io.tmpdir"), "prontuarios");
             if (!tmpDir.exists()) {
                 tmpDir.mkdir();
             }
-            File file = new File(tmpDir, fileName);
-            FileOutputStream out = new FileOutputStream(file);
+            final File file = new File(tmpDir, fileName);
+            final FileOutputStream out = new FileOutputStream(file);
             IOUtils.copy(fileInputStream, out);
             IOUtils.closeQuietly(out);
-            ProntuarioCondutorService service = new ProntuarioCondutorService();
-            return service.insertOrUpdate(file.getPath());
+            final ProntuarioCondutorService service = new ProntuarioCondutorService();
+            service.insertOrUpdate(file.getPath());
+            return Response.ok("Prontuários inseridos com sucesso");
         } catch (IOException e) {
             e.printStackTrace();
-            return Response.error("Erro ao enviar o arquivo.");
+            return Response.error("Erro ao enviar o arquivo");
         }
     }
 }