+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+### Refactors
+* Adiciona informações de fadiga no prontuário do condutor (PL-2269)
+
+
 <a name="v1.0.6"></a>
 ## Version [v1.0.6](https://github.com/luizfp/PrologWebService/compare/v1.0.5...v1.0.6) (release-date) [unreleased]